migration changes

+++ database/migrations/2025_09_12_145107_create_residential_properties_table.php
@@ -11,13 +11,13 @@ return new class extends Migration
         Schema::create('residential_properties', function (Blueprint $table) {
             $table->id();
             $table->foreignId('property_id')->constrained('property_ads')->onDelete('cascade');
-            $table->tinyInteger('bedrooms')->unsigned();
+            $table->integer('bedrooms')->unsigned();
             $table->decimal('bathrooms', 3, 1)->unsigned()->nullable();
             $table->decimal('floor_area', 10, 2)->nullable();
             $table->decimal('lot_size', 10, 2)->nullable();
             $table->year('year_built')->nullable();
-            $table->tinyInteger('floors')->unsigned()->nullable();
-            $table->tinyInteger('parking_spaces')->unsigned()->nullable();
+            $table->integer('floors')->unsigned()->nullable();
+            $table->integer('parking_spaces')->unsigned()->nullable();
             $table->boolean('balcony')->default(false);
             $table->string('heating_cooling', 100)->nullable();
             $table->json('amenities')->nullable();
